Fixing some issues with the AggregateSettingsProvider
The settings provider didn't work if one of the providers returned null for a setting, but the return type wasn't nullable. To sort this I've added a new method to the `ISettingsProvider` interface to return the unconverted string value, which allows the aggregate provider to get the value and do the conversion separately.